Minor refactor and add new grammar.txt file

+++ BirdLang/include/Cursor.h
@@ -5,14 +5,14 @@
 class Cursor
 {
 public:
-	Cursor(int index = -1, int line = 0, int column = -1, const std::string& filename = "", const std::string& input = "") :
+	Cursor(size_t index = -1, int line = 0, int column = -1, const std::string& filename = "", const std::string& input = "") :
 		index(index),
 		line(line),
 		column(column),
 		filename(filename),
 		input(input) {}
 
-	int index;
+	size_t index;
 	int line;
 	int column;
 	std::string filename;
